feat: implement rearrangeFeed mutation logic
This commit implements the core logic for the `rearrangeFeed` mutation, enabling users to reorder their feeds.

The changes include:

- Implementing the `RearrangeFeed` resolver in `backend/gql/resolver/mutation.resolvers.go`.
- Adding transaction logic to ensure atomicity of index updates.
- Implementing index shifting logic to accommodate feed reordering.
- Updating `SelectFeedMaxIdx` to return an integer instead of an interface.

+++ backend/query/feed.sql
@@ -64,7 +64,7 @@ FOR UPDATE;
 
 -- name: SelectFeedMaxIdx :one
 SELECT
-    MAX(idx)
+    MAX(idx)::int
 FROM
     feeds;
